email: move timesegmentor to models package
